+++ CHANGELOG.md
@@ -25,6 +25,7 @@ _What's new_
 * Convert credential for microsoft teams to connection. ([#474](https://github.com/turbot/pipe-fittings/issues/474)).
 * Convert credential for okta to connection. ([#475](https://github.com/turbot/pipe-fittings/issues/475)).
 * Convert credential for openai to connection. ([#476](https://github.com/turbot/pipe-fittings/issues/476)).
+* Convert credential for opsgenie to connection. ([#477](https://github.com/turbot/pipe-fittings/issues/477)).
 
 _Bug fixes_